requirements.txt
Alembic==1.12.1
# blinker==1.6.3
# click==8.1.7
Flask==3.0.0
Flask-admin==1.6.1
Flask-login==0.6.2
Flask-mail==0.9.1
Python-dotenv==1.0.0
Flask-SQLAlchemy==3.1.1
# greenlet==3.0.0
# itsdangerous==2.1.2
# Jinja2==3.1.2
# Mako==1.3.0
# MarkupSafe==2.1.3
Psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
# typing_extensions==4.8.0
Flask-WTF==1.2.1
WTForms==3.1.1
Email-validator==2.1.0
Flask-uploads==0.2.1